+++ projects/jrng-ui/src/public-api.ts
@@ -122,6 +122,7 @@ export * from 'jrng-ui/inplace';
 export * from 'jrng-ui/anchor';
 export * from 'jrng-ui/affix';
 export * from 'jrng-ui/watermark';
+export * from 'jrng-ui/speed-dial';
 export * from 'jrng-ui/transfer-list';
 export * from 'jrng-ui/tree';
 export * from 'jrng-ui/tree-table';